feat: creating error handling for artistServices

+++ src/domains/artists/services/artistsServices.ts
@@ -1,35 +1,84 @@
 import { PrismaClient } from "@prisma/client";
 import { Artista } from "@prisma/client";
+import { InvalidParamError } from "../../../../errors/errors/InvalidParamError";
+import { QueryError } from "../../../../errors/errors/QueryError";
 
 const prisma = new PrismaClient();
+
+//ToDo: Tratamento de Erros (adicionar os res.status em artists/index.ts)
 
 class ArtistService {
 	//Método para criar um novo artista
 	async createArtist(nome: string, foto?: string, numero_streams?: number) {
-		return prisma.artista.create({
-			data: { nome, foto, numero_streams },
-		});
+		try {
+			if (nome == null){
+				throw new InvalidParamError("Nome do artista não informado!");
+			}
+	
+			if (numero_streams != undefined && numero_streams < 0){
+				throw new QueryError("O número de stream não pode ser negativo!");
+			}
+	
+			return await prisma.artista.create({
+				data: { nome, foto, numero_streams },
+			});
+
+		} catch (error) {
+			console.error("Erro ao criar o artista:", error);
+			throw new Error("Não foi possível criar o artista!");
+		}
 	}
+
 	//Método para buscar um artista pelo ID
 	async getArtistById(id: number) {
-		return prisma.artista.findUnique({
-			where: { ID_Artista: id },
-			include: { musicas: true },
-		});
+		try {
+
+			if (!id){
+				throw new InvalidParamError("ID inválido!");
+			}
+
+			const artista = await prisma.artista.findUnique({
+				where: { ID_Artista: id },
+				include: { musicas: true },
+			});
+
+			if (!artista) {
+				throw new QueryError("Artista não encontrado!");
+			}
+		} catch (error) {
+			console.error("Erro ao buscar o artista", error);
+			throw new Error("Não foi possível achar o artista!");
+		}
 	}
+
 	//Método para buscar todos os artistas cadastrados
 	async getAllArtists() {
-		return prisma.artista.findMany();
+
+		try{
+			return await prisma.artista.findMany();
+		} catch(error){
+			console.error("Erro ao buscar os artistas", error);
+			throw new Error("Não foi possível buscar os artistas!");
+		}
 	}
 
 	async updateArtista(id: number, body: Partial<Artista>) {
 		try {
+
+			if (!id){
+				throw new InvalidParamError("ID inválido!");
+			}
+
 			const artistaFound = await prisma.artista.findUnique({
 				where: { ID_Artista: id },
 			});
 
 			if (!artistaFound) {
-				throw new Error("Artista não encontrado.");
+				throw new QueryError("Artista não encontrado!");
+			}
+
+			if (!body || typeof body !== "object" || Object.keys(body).length === 0) {
+				throw new InvalidParamError("Nenhum dado informado para atualização!");
 			}
 
 			const updatedArtista = await prisma.artista.update({
@@ -50,9 +99,24 @@ class ArtistService {
 
 	async deleteArtista(id: number) {
 		try {
+
+			if (!id){
+				throw new InvalidParamError("ID inválido!");
+			}
+
+			const artista = await prisma.artista.findUnique({
+				where: { ID_Artista: id },
+				include: { musicas: true },
+			});
+
+			if (!artista) {
+				throw new QueryError("Artista não encontrado!");
+			}
+
 			await prisma.artista.delete({
 				where: { ID_Artista: id },
 			});
+			
 			console.log("Artista deletado com sucesso.");
 		} catch (error) {
 			console.error("Erro ao deletar artista:", error);
